[bashrc] Some minor attempts at zsh-compatibility
So some of this can be split out to a shared-rc between bash and zsh
at some point.  With this you can source the file in zsh, but some
stuff is surely broken.  Probably needs a careful audit for subtle
bash/zsh differences, actually this is all a bad idea.